leetcode/leetcode-09/src/main/java/Solution875.java
public class Solution875 {
    public int minEatingSpeed(int[] piles, int h) {
        int left = 1;
        int right = Integer.MAX_VALUE;
        while (left < right) {
            int mid = left + (right - left) / 2;
            // 边界二分 F, F,..., F, [T, T,..., T]
            // ----------------------^
            if (checkMid(piles, h, mid)) {
                right = mid;
            } else {
                left = mid + 1;
            }
        }
        return left;
    }

    // 吃香蕉速度为 mid 时，能否在 h 小时内吃完 FFFTTT
    private boolean checkMid(int[] piles, int h, int mid) {
        int cnt = 0;
        for (int pile : piles) {
            // 数论向上取整法
            cnt += (pile + mid - 1) / mid;
        }
        return cnt <= h;
    }
}
/*
875. 爱吃香蕉的珂珂
https://leetcode.cn/problems/koko-eating-bananas/

第 94 场周赛 T3。

珂珂喜欢吃香蕉。这里有 N 堆香蕉，第 i 堆中有 piles[i] 根香蕉。警卫已经离开了，将在 H 小时后回来。
珂珂可以决定她吃香蕉的速度 K （单位：根/小时）。每个小时，她将会选择一堆香蕉，从中吃掉 K 根。如果这堆香蕉少于 K 根，
她将吃掉这堆的所有香蕉，然后这一小时内不会再吃更多的香蕉。
珂珂喜欢慢慢吃，但仍然想在警卫回来前吃掉所有的香蕉。
返回她可以在 H 小时内吃掉所有香蕉的最小速度 K（K 为整数）。
提示：
1 <= piles.length <= 10^4
piles.length <= h <= 10^9
1 <= piles[i] <= 10^9

二分
 */